+++ src/Core/Core/Execution/ExecutionStrategyBase.Resolver.cs
@@ -98,7 +98,7 @@ namespace HotChocolate.Execution
         }
 
         protected static void CompleteValue(
-            FieldValueCompletionContext completionContext)
+            IFieldValueCompletionContext completionContext)
         {
             _fieldValueCompleter.CompleteValue(completionContext);
         }