+++ src/Core/Core/Execution/SubscriptionExecutionStrategy.cs
@@ -14,7 +14,7 @@ namespace HotChocolate.Execution
     internal sealed class SubscriptionExecutionStrategy
         : ExecutionStrategyBase
     {
-        private IRequestTimeoutOptionsAccessor _options;
+        private readonly IRequestTimeoutOptionsAccessor _options;
 
         public SubscriptionExecutionStrategy(
             IRequestTimeoutOptionsAccessor options)
@@ -97,7 +97,7 @@ namespace HotChocolate.Execution
 
         private static Task<IEventStream> SubscribeAsync(
             IServiceProvider services,
-            EventDescription @event)
+            IEventDescription @event)
         {
             var eventRegistry = (IEventRegistry)services
                 .GetService(typeof(IEventRegistry));