+++ CHANGELOG.md
@@ -27,6 +27,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added schema stitching capabilities [#341](https://github.com/ChilliCream/hotchocolate/issues/341).
 - Added generic interface type [#546](https://github.com/ChilliCream/hotchocolate/issues/546).
 - Added directives support for input objects [#548](https://github.com/ChilliCream/hotchocolate/issues/548).
+- Added optional totalCount field to the connection type [#558](https://github.com/ChilliCream/hotchocolate/issues/558).
+- Added support for dynamic generated schema types [#558](https://github.com/ChilliCream/hotchocolate/issues/558).
+- Added generic union type [#552](https://github.com/ChilliCream/hotchocolate/issues/552).
+- Added options to AddStitchedSchema [#556](https://github.com/ChilliCream/hotchocolate/issues/556).
 
 ### Changed
 
@@ -37,6 +41,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The DataLoader API is now offering a simpler interface.
 - Extended Scalar Types must now be explicitly registered during schema configuration. [#433](https://github.com/ChilliCream/hotchocolate/issues/433)
 - Authorization directive is now repeatable and can use the default authorization policy [#485](https://github.com/ChilliCream/hotchocolate/pull/485)
+- UsePaging can now be used without specifying the clr type [#558](https://github.com/ChilliCream/hotchocolate/issues/558).
+- The edge node type can now be any output type including list [#558](https://github.com/ChilliCream/hotchocolate/issues/558).
 
 ### Fixed